requirements.txt
alembic==1.8.1
GeoAlchemy2==0.10.2
geopandas==0.10.2
Flask==2.1.2
psycopg2-binary==2.9.3
Pillow==9.5.0
pytest==7.1.2
python-dotenv==0.21.0
python-magic==0.4.27
SQLAlchemy==1.4.36
waitress==2.1.1
Werkzeug==2.0.3
numpy==1.23.3
Shapely==1.8.2
chardet==5.0.0
gdal==3.2.2.1